+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Features
 
+- Profile envelopes are sent directly from profiler ([#2298](https://github.com/getsentry/sentry-java/pull/2298))
 - Add captureProfile method to hub and client ([#2290](https://github.com/getsentry/sentry-java/pull/2290))
 
 ## 6.5.0